Small change in PolyhedronMesh

+++ src/collision/PolyhedronMesh.cpp
@@ -199,7 +199,7 @@ void PolyhedronMesh::computeCentroid() {
         mCentroid += getVertex(v);
     }
 
-    mCentroid /= getNbVertices();
+    mCentroid /= static_cast<decimal>(getNbVertices());
 }
 
 // Compute and return the area of a face
